python_backend/requirements.txt
python-dotenv>=0.19.0
anthropic>=0.3.0